+++ tests/python/test_parser.py
@@ -4,7 +4,6 @@
 
 from io import StringIO
 import sys
-import unittest
 from textwrap import dedent
 
 import astroid
@@ -13,7 +12,7 @@ import pytest
 from autoapi.mappers.python.parser import Parser
 
 
-class PythonParserTests(unittest.TestCase):
+class TestPythonParser:
     def parse(self, source):
         node = astroid.extract_node(source)
         return Parser().parse(node)
@@ -24,16 +23,16 @@ class PythonParserTests(unittest.TestCase):
             pass
         """
         data = self.parse(source)[0]
-        self.assertEqual(data["name"], "foo")
-        self.assertEqual(data["type"], "function")
+        assert data["name"] == "foo"
+        assert data["type"] == "function"
 
     def test_parses_all(self):
         source = """
         __all__ = ['Foo', 5.0]
         """
         data = self.parse(source)[0]
-        self.assertEqual(data["name"], "__all__")
-        self.assertEqual(data["value"], ["Foo", 5.0])
+        assert data["name"] == "__all__"
+        assert data["value"] == ["Foo", 5.0]
 
     def test_parses_all_multiline(self):
         source = """
@@ -43,27 +42,27 @@ class PythonParserTests(unittest.TestCase):
         ]
         """
         data = self.parse(source)[0]
-        self.assertEqual(data["value"], ["foo", "bar"])
+        assert data["value"] == ["foo", "bar"]
 
     def test_parses_name(self):
         source = "foo.bar"
-        self.assertEqual(self.parse(source), {})
+        assert self.parse(source) == {}
 
     def test_parses_list(self):
         name = "__all__"
         value = [1, 2, 3, 4]
         source = "{} = {}".format(name, value)
         data = self.parse(source)[0]
-        self.assertEqual(data["name"], name)
-        self.assertEqual(data["value"], value)
+        assert data["name"] == name
+        assert data["value"] == value
 
     def test_parses_nested_list(self):
         name = "__all__"
         value = [[1, 2], [3, 4]]
         source = "{} = {}".format(name, value)
         data = self.parse(source)[0]
-        self.assertEqual(data["name"], name)
-        self.assertEqual(data["value"], value)
+        assert data["name"] == name
+        assert data["value"] == value
 
     def test_arguments(self):
         """Argument parsing of source"""
@@ -74,17 +73,15 @@ class PythonParserTests(unittest.TestCase):
             "    return True\n"
         )
         data = self.parse(source)[0]
-        self.assertEqual(
-            data["args"],
-            [
-                (None, "self", None, None),
-                (None, "bar", None, None),
-                (None, "baz", None, "42"),
-                (None, "foo", None, "True"),
-                ("*", "args", None, None),
-                ("**", "kwargs", None, None),
-            ],
-        )
+        expected = [
+            (None, "self", None, None),
+            (None, "bar", None, None),
+            (None, "baz", None, "42"),
+            (None, "foo", None, "True"),
+            ("*", "args", None, None),
+            ("**", "kwargs", None, None),
+        ]
+        assert data["args"] == expected
 
     def test_advanced_arguments(self):
         """Advanced argument parsing"""
@@ -96,24 +93,22 @@ class PythonParserTests(unittest.TestCase):
             "    return True\n"
         )
         data = self.parse(source)[0]
-        self.assertEqual(
-            data["args"],
-            [
-                (None, "self", None, None),
-                (None, "a", None, None),
-                (None, "b", None, None),
-                (None, "c", None, "42"),
-                (None, "d", None, "'string'"),
-                (None, "e", None, "(1, 2)"),
-                (None, "f", None, "{'a': True}"),
-                (None, "g", None, "None"),
-                (None, "h", None, "[1, 2, 3, 4]"),
-                (None, "i", None, "dict(a=True)"),
-                (None, "j", None, "False"),
-                ("*", "args", None, None),
-                ("**", "kwargs", None, None),
-            ],
-        )
+        expected = [
+            (None, "self", None, None),
+            (None, "a", None, None),
+            (None, "b", None, None),
+            (None, "c", None, "42"),
+            (None, "d", None, "'string'"),
+            (None, "e", None, "(1, 2)"),
+            (None, "f", None, "{'a': True}"),
+            (None, "g", None, "None"),
+            (None, "h", None, "[1, 2, 3, 4]"),
+            (None, "i", None, "dict(a=True)"),
+            (None, "j", None, "False"),
+            ("*", "args", None, None),
+            ("**", "kwargs", None, None),
+        ]
+        assert data["args"] == expected
 
     def test_dict_key_assignment(self):
         """Ignore assignment to dictionary entries."""
@@ -124,7 +119,7 @@ class PythonParserTests(unittest.TestCase):
         MY_DICT['key2'] = 'value2'
         """
         data = self.parse(source)[0]
-        self.assertEqual(data["name"], "MY_DICT")
+        assert data["name"] == "MY_DICT"
 
     def test_list_index_assignment(self):
         """Ignore assignment to indexes."""
@@ -135,4 +130,4 @@ class PythonParserTests(unittest.TestCase):
         COLOUR[2] = 255
         """
         data = self.parse(source)[0]
-        self.assertEqual(data["name"], "COLOUR")
+        assert data["name"] == "COLOUR"